% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveICA.R
\name{waveica}
\alias{waveica}
\title{WaveICA}
\usage{
waveica(
  data,
  wf = "haar",
  batch,
  factorization = "stICA",
  group = NULL,
  K = 20,
  t = 0.05,
  t2 = 0.05,
  alpha = 0
)
}
\arguments{
\item{data}{Dataframe. Feature table with intensities.}

\item{wf}{String. Wavelet function, the default is "haar".}

\item{batch}{Vector. Batch number of each sample.}

\item{factorization}{String. Matrix factorization method, options are ["stICA", "SVD"]. The default is "stICA".}

\item{group}{Vector, optional. Type of a sample (blank, sample, QC) numerically encoded to blank:0, sample:1, QC:2.}

\item{K}{Integer. The maximal number of independent components (for ICA) or singular vectors (SVD). The default is 20.}

\item{t}{Float between 0 and 1. The threshold to consider a component associate with the batch. The default is 0.05.}

\item{t2}{Float between 0 and 1. The threshold to consider a component associate with the group. The default is 0.05.}

\item{alpha}{Float between 0 and 1. The trade-off value between the independence of samples and those
of variables. The default is 0.}
}
\value{
Dataframe. Feature table with intensities corrected of batch effects.
}
\description{
Removing batch effects for metabolomics data.
}
\author{
Kui Deng
\email{dengkui_stat@163.com}

RECETOX
\email{GalaxyToolsDevelopmentandDeployment@space.muni.cz}
}
