% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveICA.R
\name{waveica_nonbatchwise}
\alias{waveica_nonbatchwise}
\title{WaveICA_nonbatchwise}
\usage{
waveica_nonbatchwise(
  data,
  wf = "haar",
  injection_order,
  alpha = 0,
  cutoff,
  K = 20
)
}
\arguments{
\item{data}{Dataframe. Feature table with intensities.}

\item{wf}{String. Wavelet function. The default is "haar".}

\item{injection_order}{Vector. Injection order of the samples.}

\item{alpha}{Float between 0 and 1. The trade-off value between the independence of samples and those
of variables. The default is 0.}

\item{cutoff}{Float between 0 and 1. The threshold of the variation explained by the injection order for
independent components.}

\item{K}{Integer. The maximal number of independent components. The default is 10.}
}
\value{
Dataframe. Feature table with intensities corrected of batch effects.
}
\description{
Removing batch effects for metabolomics data without batch information.
}
\author{
Kui Deng
\email{dengkui_stat@163.com}

RECETOX
\email{GalaxyToolsDevelopmentandDeployment@space.muni.cz}
}
